requirements.txt
skyfield>=1.52.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
pyyaml>=6.0.0
plotly>=5.10.0
astropy>=5.1.0
geopy>=2.3.0
ipython>=8.0.0
dash>=2.8.0
